Test.Mutaskell

Description

MuCheck base module

Documentation

mucheck

Arguments

:: (Show b, Summarizable b, TRun a b)

=> a
The module we are mutating

-> FilePath
The HPC coverage.tix file

-> IO (Either String (MAnalysisSummary, [MutantSummary]))
Returns a tuple of full summary, and individual mutant results.

Perform mutation analysis using any of the test frameworks that support Summarizable (essentially, after running it on haskell, we should be able to distinguish a successful run without failures from one with failures.) E.g. using the mucheck-quickcheck adapter

tFn :: Mutant -> TestStr -> InterpreterOutput QuickCheckSummary`
tFn = testSummary
mucheck tFn "Examples/QuickCheckTest.hs" ["quickCheckResult revProp"]

sampler

Arguments

:: Config
Configuration

-> [Mutant]
The original list of mutation operators

-> IO [Mutant]
Returns the sampled mutation operators

Wrapper around sampleF that returns correct sampling ratios according to configuration passed. TODO: Actually use the sampling configuration.