Test.Mutaskell.AnalysisSummary

Description

The AnalysisSummary declares the mutation result datatype, and its instances.

Documentation

data MAnalysisSummary

Datatype to hold results of the entire run

Constructors

MAnalysisSummary

Fields

_maCoveredNumMutants :: Int

The number of mutants that was covered by test cases

_maNumMutants :: Int

The number of mutants tested (after sampling)

_maAlive :: Int

The number of mutants that were alive after the mutation run

_maKilled :: Int

The number of mutants that were killed

_maErrors :: Int

The number of mutants that produced interpreter errors

_maSkipped :: Int

The number of non-compilable mutants (WontCompile errors)

summaryTotal :: MAnalysisSummary -> Int

Total mutant count acting as the basis for percentage calculations (evaluated mutants).

summaryNoErrors :: MAnalysisSummary -> Int

Number of non-error mutants evaluated.

summaryMsi :: MAnalysisSummary -> Int

Mutation Score Indicator (MSI) as a percentage (0–100).

summaryCoveredNoErrors :: MAnalysisSummary -> Maybe Int

Covered mutant count excluding interpreter errors, when coverage is present.

summaryCoveredMsi :: MAnalysisSummary -> Maybe Int

Covered MSI percentage when coverage is present.

maSummary :: MAnalysisSummary -> (Int, Int, Int, Int, Int, Int)

MAnalysisSummary to tuple