Test.Mutaskell.Interpreter

Description

The Interpreter module is responible for invoking the Hint interpreter to evaluate mutants.

Documentation

evaluateMutants

Arguments

:: (Show b, Summarizable b, TRun a b)

=> Int
Number of parallel worker processes

-> Maybe Int
Optional timeout in microseconds

-> Maybe FilePath
Optional directory to keep mutant files in (Nothing = system temp, deleted after)

-> [String]
Extra arguments forwarded to every test invocation via withArgs

-> Maybe (MutantSummary -> IO ())
Optional per-mutant callback invoked after each mutant is evaluated

-> a
The module to be evaluated

-> [Mutant]
The mutants to be evaluated

-> [TestStr]
The tests to be used for analysis

-> IO (MAnalysisSummary, [MutantSummary])
Returns a tuple of full run summary and individual mutant summary

Given the list of tests suites to check, run the test suite on mutants.

evalMethod

Arguments

:: (MonadInterpreter m, Typeable t)

=> String
The mutant _file_ to load

-> TestStr
The test to be run

-> m t
Returns the monadic computation to be run by I.runInterpreter

Given the filename, modulename, test to evaluate, evaluate, and return result as a pair.

t = I.runInterpreter (evalMethod
       "Examples/QuickCheckTest.hs"
       "quickCheckResult idEmp")

evalMutant

Arguments

:: (Typeable t, Summarizable t)

=> Maybe Int
Optional timeout

-> Bool
Whether to delete the mutant file after evaluation

-> FilePath
Directory to write the mutant file into

-> [String]
Extra arguments forwarded to every test invocation

-> [TestStr]
The tests to be used

-> Mutant
Mutant being tested

-> IO [InterpreterOutput t]
Returns the result of test runs

Run all tests on a mutant

evalTest

Arguments

:: (Typeable a, Summarizable a)

=> Maybe Int
Optional timeout in microseconds

-> [String]
Extra arguments forwarded to the test via withArgs

-> String
The mutant _file_ that we have to evaluate (_not_ the content)

-> String
The file where we will write the stdout and stderr during the run.

-> TestStr
The test to be run

-> IO (InterpreterOutput a)
Returns the output of given test run

Run one single test on a mutant

summarizeResults

Arguments

:: (Summarizable s, TRun a s)

=> a
The module to be evaluated

-> [TestStr]
Tests we used to run analysis

-> (Mutant, [InterpreterOutput s])
The mutant and its corresponding output of test runs.

-> MutantSummary
Returns a summary of the run for the mutant

The summarizeResults function evaluates the results of a test run using the supplied isSuccess and summarize_ functions from the adapters

summaryFromMutantSummaries :: [MutantSummary] -> MAnalysisSummary

Build an MAnalysisSummary directly from a list of per-mutant results. Use this when results were collected outside the normal evaluateMutants call (e.g. from parallel worker subprocesses).

data MutantSummary

Data type to hold results of a single test execution

Constructors

MSumError Mutant String [Summary]
Interpreter or runtime error

MSumSkipped Mutant [Summary]
Non-compilable mutant (WontCompile)

MSumAlive Mutant [Summary]
The mutant was alive

MSumKilled Mutant [Summary]
The mutant was killed

MSumOther Mutant [Summary]
Undetermined - we will treat it as killed as it is not a success.

isSkippedSummary :: MutantSummary -> Bool

True when the summary represents a non-compilable (skipped) mutant.