Test.Mutaskell.Utils.Common

Description

Common functions used by MuCheck

Documentation

choose :: [a] -> Int -> [[a]]

The choose function generates subsets of a given size

coupling :: Eq a => (a -> a -> t) -> [a] -> [t]

The coupling function produces all possible pairings, and applies the given function to each

replaceFst :: Eq a => (a, a) -> [a] -> [a]

The replaceFst function replaces first matching element in a list given old and new values as a pair

sample :: RandomGen g => g -> Int -> [t] -> [t]

The sample function takes a random generator and chooses a random sample subset of given size.

rSample :: Int -> [t] -> IO [t]

Wrapper around sample providing the random seed

sampleF :: RandomGen g => g -> Rational -> [t] -> [t]

The sampleF function takes a random generator, and a fraction and returns subset of size given by fraction

rSampleF :: Rational -> [t] -> IO [t]

Wrapper around sampleF providing the random seed

remElt :: Int -> [a] -> [a]

The remElt function removes element at index specified from a list

swapElts :: Int -> Int -> [t] -> [t]

The swapElts function swaps two elements in a list given their indices

genSwapped :: [t] -> [[t]]

The genSwapped generates a list of lists where each element has been swapped by another

genRandomSeed :: IO StdGen

Generate a random seed from the time.

curryM :: (t1 -> t2 -> m t) -> (t1, t2) -> m t

take a function of two args producing a monadic result, and apply it to a pair

hash :: String -> String

A simple hash

spread :: (a, [b]) -> [(a, b)]

convert a tuple with element and second array to an array of tuples by repeating the first element

apSnd :: (b -> c) -> (a, b) -> (a, c)

Apply a function to the last of a tuple

apTh :: (c -> d) -> (a, b, c) -> (a, b, d)

Apply a function to the last of a 3-tuple

strip :: String -> String

Strip whitespace from ends