Test.Mutaskell.AnalysisSummary

Description

The AnalysisSummary declares the mutation result datatype, and its instances.

Documentation

data MAnalysisSummary

Datatype to hold results of the entire run

Constructors

MAnalysisSummary

Fields

_maCoveredNumMutants :: Int

The number of mutants that was covered by test cases

_maNumMutants :: Int

The number of mutants tested (after sampling)

_maAlive :: Int

The number of mutants that were alive after the mutation run

_maKilled :: Int

The number of mutants that were killed

_maErrors :: Int

The number of mutants that produced interpreter errors

_maSkipped :: Int

The number of non-compilable mutants (WontCompile errors)

maSummary :: MAnalysisSummary -> (Int, Int, Int, Int, Int, Int)

MAnalysisSummary to tuple