Test.Mutaskell.Tix

Description

Read the HPC Tix and Mix files.

Documentation

type Span = HpcPos

Span info - same as HpcPos

toSpan :: (Int, Int, Int, Int) -> Span

Convert a 4-tuple to a span

spanStartLine :: Span -> Int

Extract the 1-based start line from a span.

data TCovered

Whether a line is covered or not

Constructors

TCovered

TNotCovered

isCovered :: TCovered -> Bool

Whether a line is covered or not

insideSpan :: Span -> Span -> Bool

insideSpan small big

mixTix :: String -> Mix -> TixModule -> (String, [(Span, TCovered)])

mixTix joins together the location and coverage data.

parseTix :: String -> IO [TixModule]

reads a tix file. The tix is named for the binary run, and contains a list of modules involved.

getMix :: TixModule -> IO (Either String Mix)

Read the corresponding Mix file to a TixModule. Returns Left with a user-readable message if the .mix file cannot be found.

tryReadMix :: [FilePath] -> Either String TixModule -> IO (Maybe Mix)

Helper to try reading a mix file without crashing

getMixedTix :: String -> IO (Either String [(String, [(Span, TCovered)])])

return the tix and mix information, or a Left error if any .mix file is missing.

getUnCoveredPatches :: String -> String -> IO (Either String (Maybe [Span]))

getUnCoveredPatches returns the largest parts of the program that are not covered. Returns Left with a user-readable error if a .mix file is missing.

getNamedModule :: String -> [(String, [(Span, TCovered)])] -> [(Span, TCovered)]

Get the span and covering information of the given module

removeRedundantSpans :: [Span] -> [Span]

Remove spans which are contained within others of same kind.